+++ src/components/Tooltip/Tooltip.styles.css
@@ -1,0 +1,1 @@
+@import "../../index.css";
