+++ src/index.js
@@ -15,3 +15,4 @@ export { default as Spinner } from './components/Spinner/Spinner.component';
 export { default as CountdownTimer } from './components/CountdownTimer/CountdownTimer.component';
 export { default as HalfPieChart } from './components/HalfPieChart/HalfPieChart.component';
 export { default as LineChart } from './components/LineChart/LineChart.component';
+export { default as Tooltip } from './components/Tooltip/Tooltip.component';
